+++ src/main.rs
@@ -1,4 +1,6 @@
 extern crate clap;
+extern crate regex;
+
 use clap::{App, SubCommand};
 
 mod tasks;
@@ -15,10 +17,20 @@ fn main() {
         .subcommand(SubCommand::with_name("sync").about(
             "install / update my settings on this computer",
         ))
+        .subcommand(SubCommand::with_name("update").about(
+            "update packages on this computer",
+        ))
         .get_matches();
 
     if let Some(_matches) = matches.subcommand_matches("sync") {
         println!("syncing...");
         tasks::sync();
+        return;
+    }
+
+    if let Some(_matches) = matches.subcommand_matches("update") {
+        println!("updating...");
+        tasks::update();
+        return;
     }
 }