+++ src/tasks/mod.rs
@@ -1,6 +1,7 @@
 mod alacritty;
 mod hyper;
 mod psql;
+mod rust;
 mod ssh;
 mod tmux;
 mod vim;
@@ -10,8 +11,20 @@ pub fn sync() {
     alacritty::sync();
     hyper::sync();
     psql::sync();
+    rust::sync();
     ssh::sync();
     tmux::sync();
     vim::sync();
     vscode::sync();
+}
+
+pub fn update() {
+    alacritty::update();
+    hyper::update();
+    psql::update();
+    rust::update();
+    ssh::update();
+    tmux::update();
+    vim::update();
+    vscode::update();
 }