+++ CHANGELOG.md
@@ -7,9 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* add `update` command
+
+* implement `update` to update Rust with `rustup`
+
+* implement `update` to update crates installed by `cargo install`
+
+### Changed
+
 * use values from Cargo.toml for `--help` and `--version`
 
 ## [0.1.2] - 2018-02-18
+
+### Changed
 
 * add missing package metadata to fix `cargo publish`